DDBTEAM-603: Updated VendorCoreService inject and trait usage
