+++ config/services.yaml
@@ -41,6 +41,8 @@ services:
     _instanceof:
         App\Service\VendorService\VendorServiceInterface:
             tags: [app.vendor.service]
+            calls:
+                - setVendorCoreService: ['@App\Service\VendorService\VendorCoreService']
 
     # makes classes in src/ available to be used as services
     # this creates a service per class whose id is the fully-qualified class name
